+++ src/apps/hadolint.rs
@@ -1,0 +1,43 @@
+use crate::apps::{GetRTACmdArgs, get_rta_command};
+use crate::config::{Application, ApplicationSection, Config, Operation};
+use crate::domain::{DetectedStack, EnabledWhen, Lint, Result, Tool};
+use std::fmt::Display;
+
+pub struct Hadolint;
+
+impl Tool for Hadolint {
+    fn enabled_when(&self) -> EnabledWhen {
+        EnabledWhen::Always
+    }
+
+    fn config_section<'a>(&self, apps: &'a ApplicationSection) -> Option<&'a dyn Application> {
+        Some(apps.hadolint.as_ref()?)
+    }
+}
+
+impl Display for Hadolint {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str("hadolint")
+    }
+}
+
+impl Lint for Hadolint {
+    fn lint_commands(
+        &self,
+        stack: &DetectedStack,
+        config: &Config,
+    ) -> Result<Option<conc::Runnable>> {
+        let ignores = config.ignores_for(self, Operation::Lint)?;
+        let files = stack.files.remove(&ignores);
+        if files.is_empty() {
+            return Ok(None);
+        }
+        let executable = get_rta_command(&GetRTACmdArgs {
+            name: format!("lint {} ({self})", stack.stack),
+            app: &rta::applications::Hadolint {},
+            args: files.into_strings(),
+            version: None,
+        })?;
+        Ok(executable.map(conc::Runnable::Single))
+    }
+}